pyproject.toml
[project]
name = "challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "ipython (>=8.31.0,<9.0.0)",
    "django-summernote (>=0.8.20.0,<0.9.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "django-restframework (>=0.0.1,<0.0.2)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
